Create HTML Code
This code is a simplified structure of an HTML webpage used to display the basic layout of an online bookstore. It includes a header (title, character set, viewport settings), body (welcome message, navigation menu, book list area), and footer (copyright information). Use CSS for style design to achieve responsive layout that adapts to different device screen sizes. The book information is a placeholder and needs to be added in practice.